+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `CrowdfundingCard` mobile version.
 - Feature: Delete `stateDay` and update `dayCounter` props on `CrowdfundingCardWidget`
 
 ## [2.38.0] - 2019-11-08